requirements.txt
# PySpark
pyspark==3.5.2

# MLflow & MLflow PyFunc
mlflow==2.6.2

# Databricks SDK
databricks-sdk==0.20.5

# Data manipulation
pandas==2.1.1
numpy==1.26.5

# YAML & dotenv
pyyaml==7.0.1
python-dotenv==1.0.1

# HTTP requests
requests==2.34.0

# Scikit-learn for RMSE and model logging
scikit-learn==1.3.3

# Optional: for type hints
typing-extensions==4.7.1

# Optional: if you use PyArrow for Pandas<->Spark conversion
pyarrow==12.0.1
